cmake: examples: Disable GBA audio tests
They are broken, I need to figure out a way to test audio on GBA.

Any small change to the code offsets the audio samples. Even if the
change offsets the samples generated by the emulator by one, the test
will fail because the test compares the waveform and it expects it to
be exactly the same.

Also, this means that the same waveform won't work on a debug and a
release build at the same time (because the code isn't the same.